dsp: q6asm: fix consecutive cal_memory unmap failure
buffer_node for cal_memory are maintained in asm
commont_client IN-port lists.
Current codes work fine for the pattern:
map, unmap, map, unmap, map, unmap...
But it fails at the second unmap in the cases:
map, map, map, unmap, unmap, unmap, since the
buffer address in the common_client has not been
updated and we are actually trying to unmap
the same buffer multi-times.

Change-Id: I51f442dc8063a55932f44b8b7f3c689029565f7b
Signed-off-by: Guodong Hu <guodhu@codeaurora.org>